requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
psycopg2-binary
matplotlib
pandas
streamlit
sqlalchemy
pymongo